+++ src/cli/args.rs
@@ -1,11 +1,14 @@
-use pareg::ParegRef;
+use std::ops::Range;
 
-use crate::{cli::help, err::Result};
+use pareg::{ParegRef, parsef};
 
-#[derive(Debug, Default)]
+use crate::{cli::help, err::Result, tone::Tone};
+
+#[derive(Debug)]
 pub struct Args {
     pub midi_port: Option<String>,
     pub helped: bool,
+    pub tone_range: Range<Tone>,
 }
 
 impl Args {
@@ -19,6 +22,23 @@ impl Args {
                     res.helped = true;
                 }
                 "-p" | "--port" => res.midi_port = Some(args.next_arg()?),
+                "-r" | "--range" => {
+                    let arg = args.next_arg::<&str>()?;
+                    args.map_res(parsef!(
+                        &mut arg.into(),
+                        "{}..{}",
+                        &mut res.tone_range.start,
+                        &mut res.tone_range.end
+                    ))?;
+                    if res.tone_range.start >= res.tone_range.end {
+                        return args
+                            .err_invalid()
+                            .hint(
+                                "First value must be smaller than the second.",
+                            )
+                            .err()?;
+                    }
+                }
                 _ => args.err_unknown_argument().err()?,
             }
         }
@@ -28,5 +48,15 @@ impl Args {
 
     pub fn run(&self) -> bool {
         !self.helped || self.midi_port.is_some()
+    }
+}
+
+impl Default for Args {
+    fn default() -> Self {
+        Self {
+            midi_port: None,
+            helped: false,
+            tone_range: Tone::C3..Tone::C4,
+        }
     }
 }