+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
 ## future
+- Add help when for guessing.
+- Add option to specify the range.
 - Proper CLI and errors.
 - Play tone and guess.